Updated summary-mergescore to include API in package
