+++ .gitignore
@@ -1,0 +1,8 @@
+__pycache__
+cache/
+*.pyc
+*.pyo
+*.pyd
+dist/
+build/
+*cache*/
